revert to rgb depth buffer view
It seems worthwhile trying to represent the absolute distances in the
depth visualization so we might recognise depth scale mistakes easily if
the pallet for videos we're used to seeing change. Depending on the
min/max depth range it's also possible the 256 levels of grey can't
represent the 11-16bits of depth precision we tend to have from the
device. The greyscale code is currently kept within an #if 0 guard and
maybe we could have a property to switch between the modes.