tracking: focus with small window
In several of our test videos even though the camera is generally
pointing at the person being tracked there are still times when a ray
cast from 0,0 along the Z axis may not hit the person which means our
simple clustering will fail.

We now search a small, centred, window 1/8th the size of the frame and
pick the point with the nearest depth as the starting point for the
flood fill.