tracking: use 3x3 gap filling based on random neighbour
This gap filling seems to fill in the Tango lense distortion artefacts
consistently and not otherwise interfere with segmentation / inference.

The filter is done as a 3x3 box filter that samples the 8 neighbours of
a pixel when we see a NAN depth value and then iterates the neighbours
from a random [0,7] starting point looking for a non-NAN value to fill
with.

It's aggressive in the sense that it will always fill with a non-NAN
value if any neighbours are non-NAN, which has the result of dilating
silhouettes but this doesn't seem to cause too much interference.