WIP: Allow for multiple anchor points
This change allows for multiple anchor points on the skeleton to be
specified, where an anchor point is considered to be a point about which
the skeleton will be constructed.

The up-shot is that anchor points won't be moved when trying to sanitise
bones and their positions will be considered as absolute and be
interpolated between frames.