Expose minimal gm_rdt_ training api
This starts making the training code available via a shared api and
with the train_rdt command line tool being built on that api.

As part of this change I ended up down a rat hole of signed/unsigned
type mismatch errors + found some inconsistent use of sized types. (In
part this was related to exposing state via gm_ui_properties which
support int types but but not specifically [u]int32_t)

This patch makes some broad stroke changes to avoid lots of sized types
where it seems reasonable to use an int for 99% of the misc integer
variables we have in our code where we aren't packing data tightly for
memory optimizations for for serializing (also not be fussy about using
unsigned integers for things like widths/heights considering C's foot
guns when mixing signed/unsigned arithmetic).

A minor change to the train_rdt command line interface was to drop the
'limit', 'skip' and 'shuffle' options considering the tool loads a
specified index where it's easy to control the size / shuffle of the
index instead.