tracking: more cloud debug stages + class colouring
In the viewer we can now visualize the cloud after motion based
classification. When we hit our naive segmentation path we can choose to
see the points clustered while detecting the floor and the final points
from naive segmentation.

So long as the selected stage comes after classification then the points
can be coloured based on the determined labels.

While building the debug point cloud we also track the indices of
associated pcl points which can be referenced while colouring points so
e.g. points projected into the codebook or ground-aligned coordinates
can still be textured with the video correctly.