rdt: never sample background pixels
When we generate our random sample points across our data set we now
make sure to only sample points within the silhouette of the pose.  Our
tracking assumes the body has already been segmented from the background
and so there's no need to waste capacity in the tree by learning to
explicitly classify the background.

For now (for simplicity) we leave a redundant slot in the histograms for
the background label.

Another possible benefit to this change is that we should now expend
more consistent amounts of training energy on each example pose
independent of the size of the body within the frame. I.e, previously we
might have made better classifications with larger people, or people
closer to the camera since more training samples would have overlapped
with them.

Testing via test_rdt actually showed no tangible change in classification
accuracy with this change but at the very least it seems appropriate to
aim to reduce our label count which will reduce the size of the trees we
load for runtime inference.