rdt: save training results in .json only
train_rdt now only supports saving results to a .json file and
json-to-rdt should be used to create a packed rdt file for runtime
inference.

The intention here is that train_rdt should easily be able to extend
what meta data it saves as part of its results without needing to extend
the .rdt format if it's not needed at runtime.

The code for save_tree_json() has been moved into glimpse_rdt.cc where
it can start to be extended to not only save RDTree data but also any
other ancillary data that could be useful - such as hyperparameters
specified on the command line or benchmarking statistics.

Although it'll probably make sense to remove rdt-to-json at some point,
considering that the json-to-rdt conversion will become lossy, this
patch also copies the save_tree_json() code into rdt-to-json.c for now
so we don't have to worry about breaking how it currently works.

train_rdt now prints a usage error if the output filename ends in .rdt

While updating the json-to-rdt help message this also copies some
missing --help info from json-to-rdt.py just to keep them consistent.