train_rdt: remove --background command line option
Conceptually the definition of the background label comes from the
training data and is fixed.

If we need to change the label we use for the background then we could
add something to the meta.json file at the top of our training data
directories and then gather_train_data() would forward the label to
glimpse_context.cc.

Removing the option also makes it a bit easier to consider removing
other special case handling of the background label.