train_rdt: Support queuing multiple training runs
This changes the command line interface of train_rdt so that we can load
a JSON description of multiple training runs to be queued and run
sequentially.

The configuration is generalised based on the enumerable properties
advertised by glimpse_rdt.

Some default values can be specified in case they aren't defined for the
individual runs; those defaults can be overridden, per-run via the JSON
configs and finally -p <name> -v <value> options can be given to
override properties across all runs.

The --help is quite detailed (including an example of the config schema
and a dynamically generated list of properties that can be changed) and is
hopefully comprehensive.

The brief synopsis is:
  train_rdt [options] -q <run-queue.json> [<index_name>] [<results.json>]
  train_rdt [options] -q '[{ "prop": "value", ...}, ...]' \
              [<index_name>] [<results.json>]
  train_rdt [options] <index_name> <results.json>

Where a config given via -q or --queue could look something like:

  [
    {
      // Run 0...
      "index_name": "tree0",
      "n_pixels": 2000,
      "n_thresholds": 50,
      // ...
      "out_file": "tree0.json"
    },
    {
      // Run 1...
      "index_name": "tree1",
      "n_pixels": 1000,
      "n_thresholds": 100,
      // ...
      "out_file": "tree1.json"
    }
  ]

To help avoid absolute paths in the configs, the -d,--data-dir path will
be prepended to any relative "data_dir" property.