rdt: schedule thread pool work via queue
This adds a work queue that threads can pop work from when they are idle
which sets us up to allow slightly more flexible modes of scheduling.

The addition of new work is notified via a condition variable that can
wake idle worker threads. The addition of results is also notified via a
condition variable for the main thread to be able to sleep while it's
waiting for the completion of all scheduled work.

This also avoids a deadlock seen with the barrier based synchronization
in cases where we are interrupted or run out of data to process. In
these cases then the pre-requisite number of threads weren't reaching
the barrier being waited on so the main thread would wait forever.