rdt: reset metrics per-level
Metrics are now only reported if "profile" = true

We now also report periodically on early levels that are spending a
majority of the time accumulating pixels.

The thread metrics are reset at the start of each tree level and saved
at the end of each level in preparation for saving out per-level metrics
in the final results .json.