image-pre-processor: Ensure background has a label of 0
So we can avoid having an awkward hard-coded bg_label=33 in glimpse_rdt
or an awkward user-configurable background label property which is
liable to get incorrectly specified we now assume that a label of 0 from
the pre-processors represents the background.

This also updates glimpse_rdt.cc to assume a label of 0 represents the
background.

Note that although we now use label 33 for the top-left of the head we
haven't needed up update the joint-map.json since we only define the
final head position based on the lower head labels.