indexer: seed = index name + sample with[out] replacement
By default the seed used for random sampling is the name of the index
being created otherwise if we forget to specify a seed explicitly we are
likely to create multiple index files for training separate trees that
list the same randomly sampled frames.

The default sampling strategy is now to sample with replacement.  A
-w,--without-replacement option was added so sampling without replacement
is still possible too.