rdt: fix bad bug with sampling root node pixels
The sampling first builds an array in-body pixels, then generates a
random array of indices into the first array which are sorted.

The problem was that we used the randomly generated indices to index
into the actual image, skipping the indirection through the array of
in-body pixels so we'd mostly end up sampling background pixels!