train_utils: generalize api for loading training data
The current gather_train_data() has been a convenient to have utility
but it's also a bit awkward to maintain in terms of lots of in/out
arguments and catering to multiple different use cases for accessing
the training data.

In the test_rdt tool we'd now like to access the per-frame .json data
which gather_train_data() doesn't handle for us. Instead of supporting
more conditional gathering here this starts to generalize the way we
load training data so it's practical to access only the specific
per-frame data you need.