rdt_tree: track background depth threshold
The UVT combos tracked at each node of the decision tree are sensitive
to the depth value that was used for background pixels while training so
it makes sense to track this value as part of the tree itself.

This bumps the RDTree ABI version to add a bg_depth to the header.