rdt_tree: control when we accept incomplete trees
This makes the validation while loading a decision tree from json
configurable so we can say whether it's ok to load an incomplete tree,
with some untrained nodes, or if that should be treated as an error.

When we reload a pre-existing tree to continue training with the tree
then we should accept an incomplete tree but when we are using the tree
for classification we shouldn't accept an incomplete tree.