rdt: track pedantic details about sampling implementation
To allow for fairer comparisons of the accuracy for different trees it's
important that we sample UV depth values and test gradient thresholds
in exactly the same way as the sampler used while training.

We can't assume that the latest sampler code in infer.cc will be
appropriate for an arbitrary tree and so we now track details about the
changes to sampling that should allow for our test_rdt tool to match
its sampling method to the method use while training.