infer: support old and new sampling methods
This enables infer_labels to check the sampling requirements of the
current decision tree to determine wither to sample with round-nearest
UVs and discrete millimeter units or round-floor UVs and float meter
units.

This e.g. means glimpse_viewer will be able to run with old and new
decision trees.