Replace gm_pose.valid with gm_pose.type
Allow multiple types of pose, currently a 'to start' pose and a
'to ground' pose. This is mainly to allow for poses on devices that don't
have a good idea of global positioning, but do have an accelerometer so
we can know how the ground is oriented for easier segmentation.