Add a tool to do offline motion capture from recordings
This adds a tool that can load a glimpse recording and output files and
an index that can be read in via gm_target or gm_skeleton_new_from_json.

This also introduces a change to the recording device that stops is from
sending frames when no frame request has been made, and also adds a
configuration property to disable frame-skipping.