Add ability to use flipped image in inference
Add the ability to provide a flipped mapping table to infer_labels.
This lets us increase the capacity of the tree by using nodes that
have learnt how to identify the left side of the body for the right,
and vice-versa.

It's possible that this may also allow us to add more capacity to our
training runs by only training to recognise one side of the body, but
training more combinations and/or on more pixels.