unity: enable OSX usage
In particular even though we use core gl with the OSX Unity editor,
unlike on Linux the editor doesn't use direct GL state access so we use
should use glBindTexture like we do on iOS/Android instead of
glBindTextureUnit.

Another issue with the ifdef guarding of the render hooks implies the
plugin was broken on Linux/desktop recently too.