Use the in-tree label map for horizontal flipping and update RDT format
Rather than using a separately specified label-map, used the one that's
stored in the tree when training. Also update the binary tree format to
include the flip-map (and derive the flip-map from the label-map when
using JSON trees).

WARNING: The binary tree format is updated, old rdt trees are now invalid.