Remove rdt-to-json
Converting a tree from JSON, to RDT, back to JSON, and then back to RDT
does not produce the same RDT tree anymore because of the flip-map being
derived from the label-map.

There's not really any good reason to convert a binary tree to a JSON one,
so let's remove the tool and lower maintenance burden.