ios-xcode-arm64-cross-file: specify default -isysroot
Meson will fail to sanity check the compiler if we don't specify any
-isysroot in our ios-xcode cross-file.

This is a hacky workaround for the fact that Meson doesn't let us
dynamically add compiler arguments before these sanity checks are run.
We are assuming that the -isysroot that we add within our meson.build
will supersede the hardcoded path we add to the cross-file.