recording2target: follow same event-loop design as viewer/plugin
recording2target was avoiding some boilerplate for an event loop by
handling things directly in the device/context event notification
callbacks but this didn't seem ideal for a few reasons:

Conceptually the intent of the event-based design was to help decouple
the device code from the context/tracking code, E.g. allowing either
subsystem to re-work/design how they may use threads and handle
locking/reentrancy. Handling events synchronously means we are holding
up (conceptually) unknown threads which may be holding arbitrary locks
and could risk a deadlock if we trigger another event to be delivered
that leads us back into the first subsystem to re-take a lock.

Having the design and code layout between the viewer, unity_plugin and
recording2target be as similar as possible (as far as is applicable)
should help in being able to cross-reference code between tools and in
cases where we need to fix bugs or change the design it should be easier
if the tools are as consistent as possible.

Incidentally, I think this also resolves a race condition 'FIXME' since
event processing itself is now single threaded.

This reverts a glimpse-device change to maybe_notify_frame_locked to
keep the lock around the ->event_callback as a guard against code doing
too much within an event notification callback and possibly catching
cases where events aren't processed via a mainloop as intended.

This adds a config->recording.lockstep_io option that is set to true by
recording2target when opening a recording device and has the effect of
ensuring that only one frame is every read per gm_device_request_frame()
call. Previously there was the possibility of requesting with
_DEPTH|_VIDEO flags and if it were fulfilled with a _DEPTH-only frame it
was possible that a subsequent _DEPTH|_VIDEO frame would be read and
delivered to fulfill the remaining _VIDEO flag, but before the
_DEPTH-only frame was processed.

This patch adds lots of comments to highlight more about intent behind
processing events via a mainloop.