device: rework recording playback
The code for recording playback was complex and while debugging some
problems with recording2target and struggling to follow the logic it was
a good time to try and simplify (considering I introduced some of that
complexity).

The playback loop now follows these steps:
    - Read a new frame
    - Update logical/wall clocks
    - Throttle due to fast playback
    - Throttle until a frame is requested
    - Deliver the current frame
    - Skip frames if playback slow
    - Determine next frame (loop or clamp to last frame)

This remove the need for a special inner loop for the paused state
since the last step will effectively handle this. There's no need
to have code for copying device buffers for paused frames; there
doesn't need to be anything special about re-reading the same frame
besided ensuring the clocks continue to progress forwards.