glimpse_unity: avoid overriding device timestamps
This avoids hacks to ensure all device frame timestamps were
CLOCK_MONOTONIC based so we could support a gm_unity_get_time() API.

Originally we implemented that API so we could have a way for the plugin
to derive timestamps that could be used to request tracking
interpolations for the past.

At this point unity plugin and viewer both consistently derive
timestamps based on the most recent video frame timestamps (assuming
we have a decent video framerate even if depth sampling / tracking
may be slower).