Support a stub 'null' device that doesn't capture any frames
This gives us a fallback device that can be opened without any physical
camera hardware or recording assets.

The viewer will now fallback to opening a null device e.g. if we fail to
find a Kinect camera.

The plan is to also start using this from the Unity plugin while in edit
mode, so we can still access the target poses api without any
live/realtime frame capture or tracking.