unity: avoid singleton plugin + support null device
This allows the C# code to be more explicit about initializing and
terminating plugin state with all entrypoints now explicitly taking a
plugin handle instead of assuming a global singleton.

So that we can start initializing a tracking context even in edit mode
(so we can access the target pose apis) this enables forcing use of
a null camera device when initializing the plugin state.