Remove error-correction metadata from skeleton/joint structures
This includes a rewrite of the error-correction code to be based on
deviation from the mean in tracking history, rather than the previous
mix of various ad-hoc mechanisms.

This code isn't well-tested because reasons, so bone sanitisation is
now default off until we've had a chance to properly test and tweak it.