Stop hard-coding training camera intrinsics
We now derive training camera intrinsics from depth camera resolution and
the vertical fov of the first loaded decision tree. There is a configurable
parameter to down-scale the depth cloud before inference.

TODO: Some kind of tolerance to allow for very similar intrinsics to avoid
being reprojected. Note that depth information that has intrinsics that
cause the reprojection to spread it over a wider area can end up
introducing gaps in the depth buffer used for inference, and cause
inference quality to reduce. Ideally, the loaded model would have
intrinsics that match the camera it's being used with.