indexer: support --all to keep everything after exclusions
I.e. don't use any random sampling, just write N frames to each index
being created. The plan is to also support filtering the input based on
the body type and bvh file a frame corresponds to and --all should be
useful in cases where we just want to apply exlusions and filters.