image-pre-processor: support config for what noise to apply
This allows us to maintain device specific config files under
glimpse-training-data/pre-processor-configs/ for describing the noise
that best represents the characteristics of the real device.

For example the iphone-x-config.json looks like:

    {
        "noise": [
            {
                "type": "gaussian",
                "fwtm_range_map_m": 0.02
            },
            {
                "type": "perlin",
                "freq": 0.1,
                "amplitude_m": 0.15,
                "octaves": 4
            }
        ]
    }

Previously we hard-coded adding a +/- (up to) 1cm offset based on a
gaussian distribution, which matched the kind of noise we saw for the
Zenfone AR camera. Now we can combine gaussian noise with perlin noise
of any frequency / amplitude.

Some of the image-pre-processor command line options that relate to
controlling the pre-processing itself have now become properties
that can be controlled under a "properties": {} section in this
config:

    "properties": {
        "background_depth_m": 100.0,    // The depth of background pixels
        "min_body_size_px": 3000,       // discard frame if body has fewer pixels
        "min_body_change_percent": 0.1, // discard frame if changed less than
                                        // this relative to previous frame
    },

Also see glimpse-training-data/pre-processor-configs/README.md for
details.