context: updated backbuffer scheme for label_probs buffer
This attaches a scratch 'backbuffer' vector to gm_context that can be
used across tracking stages and potentially swapped into a tracking
object if we decide we've detected a person.

This approach means that once enough storage has been allocated for a
backbuffer and N in-flight tracking objects then the storage will start
to be recycled, so it hopefully also helps avoid fragmenting the heap
over a long time.

This fixes a crash when the viewer requests an rgb_label_map image due
to the tracking->label_probs pointer possibly being NULL for a tracking
object where tracking failed to find any candidate clusters.