context: allow enumeration of joint names/semantics via context
This adds gm_context_ apis for querying the number of joints that will
be part of tracking skeletons as well as names and semantic identifiers
for each joint. These APIs are also exposed to the Unity plugin which
should be able to avoid numerous hard-coded name tables and assumptions
about semantics.