context: sweeping re-ordering of code considering dependencies
Considering that glimpse_context.cc has quite a lot going on within it
and got pretty large it had also got a bit disorganised.

This tries to order things from least to most abstract to continue
avoiding needing various function prototypes at the top of the file
for internal dependencies but to also group somewhat orthogonal
sub-components. (This should also help if we decide to split the file
up).

Overall the layout is now like this:

- trivial utilities (maths, timestamps)
- misc common utilities
- frame objects
- tracking objects
- prediction objects
- tracking pipeline stage implementations
- core tracking implementation (context_track_skeleton())
- context object
- bones
- skeletons

Maybe bones/skeletons should move up higher but for now this seems ok.

One other cleanup was to ensure we're consistent with only adding a
'gm_' prefix for exported APIs, so for example
gm_context_track_skeleton() is now context_track_skeleton() since it's
an internal implementation detail for the tracking thread.

The initial impetus for this was that the various gm_tracking_create_XYZ
apis could be made static now that they are only used internally to
create per-stage debug images but removing the corresponding prototypes
from glimpse_context.h meant they couldn't stay where they were without
a forward declaration/prototype.