context: rearrangement while reviewing state updated by tracking
The plan is to support a 'paused' state that will let us disable saving
any state after tracking so that repeated frame submissions will be
processed with the same initial state. This moves the tracking history
and pose state that's currently updated by tracking to the end of the
ctx struct.

This also renames ctx->depth_class' to 'ctx->downsampled_depth',
'ctx->depth_seg' to 'ctx->seg_codebook', makes 'reset_pose' local to the
tracking function.