viewer: draw playback props (duplicated) in draw_playback_controls
The viewer needs to have more control over pausing playback (e.g. so it
can put the context into a paused state) so we're setting things up so
that the viewer is responsible for drawing all the playback controls
grouped together even though this currently duplicates the recording
device properties that we introspect too.