context: recognised frames marked as 'paused'
We are now able to avoid making any context changes while tracking
paused frames that might affect the results of re-tracking that same
frame.

This makes it more practical to investigate problem frames especially
when it comes to the motion based segmentation since we can change
various context properties that affect tracking or simply take our
time reviewing a paused frame in the viewer without the tracking
pipeline behaving as if the scene has no motion in it (which would
slowly pollute the codebook while investigating before).