context: insert codewords sorted by mean depth + match nearest not first
Considering that the mean distance of a codeword can change over time
then it's possible codewords could drift close together and so we
shouldn't be assuming that the first entry that comes within the seg_tb
threshold is the best/closest codeword for the current depth.

When inserting new codewords we now insert sorted by the depth

When matching codewords we look for the closest match which may not be
the first match.

The update_depth_codebook() code has been moved into
stage_update_codebook_cb() and ensures that all updates will have
metrics accounted for (previously some tracking failure cases would
update the codebook but we didn't account for the time under any stage
metrics).

Some comments / #warning notes have been added for some logic that seems
confusing and warrants follow-up investigation to clarify.