context: stage_codebook_classify: avoid lots of pointer indirection
Inspecting this code it seemed like there was a lot of (easily
avoidable) pointer indirection, with a good chance that it wouldn't be
possible for the compiler to squash or hoist them out of the
inner loop.

This makes some simple changes to at least give some peace of mind that
we aren't leaving performance on the table.

Objectively I wasn't actually able to notice a significant performance
difference with this change, but subjectively I find it a bit cleaner
with these changes.