Audit and make y-axis flipping consistent
There have been lots of long-standing FIXME notes in glimpse_context.cc
that we shouldn't need to flip the Y axis of our debug geometry, and
we've long been aware that we've probably been lazy and flipped the
y-axis in odd places without careful consideration beyond visually
getting the right orientation.

The only places where we should now be flipping the y-axis are at points
where we convert from 2D coords (with y=0 at the top, extending down)
into 3D frame coordinates with +Y extending up and -Y, or vice-versa.

It's no surprise that we're getting our ground alignment in a tangle
without a consistent view of which way is UP :)