Track per-frame gravity vector if supported by device
It's convenient/practical to keep track of/record a frame's orientation
according to a gravity vector considering how easy it is to visualize
as part of the tracking pipeline to sanity check what we're dealing
with.

For now this doesn't remove tracking a quaternion as part of a pose
and that's still what we use to build the transform while tracking but
visualizing the gravity vector has clearly proven that our original
gravity vector wasn't aligned with gravity as we thought.

The gravity values we get from CoreMotion are defined relative to a
portrait view of a phone, looking at the display, with +X to the right,
+Y up and +Z towards the viewer. We are careful to flip the X-axis now
(since the camera is facing the viewer) and we also swizzle the X/Y axis
based on the physical rotation of the device so that +Y up, always
corresponds to the up direction of the *rotated* frame.