context: add edge detection/deletion stage (enabled by default)
This adds an optional edge detection stage while tracking (after
downsampling) which is aims to remove the points that trail/smear off
the edges of foreground objects, often making it look like the
foreground is connected to the background.

For each point we're comparing the vector from the camera origin to that
point with the gradient vector from the point to it's immediate
neighbours.  If those two vectors are coincident we consider that to be
an edge.

This definition is specifically geared towards the kinds of trailing
artefacts that we see with the iPhone TrueDepth + dual camera data.

The code is strictly only looking at either horizontal or vertical
gradients, only looking at two points at a time. We're either looking at
the left/right neighbours and ignoring the y-components of all vectors,
or we're looking at the above/below neighbours and ignoring the
x-components of all vectors.

Note: basically nothing has been done to optimize the implementation of
this stage so far, and there's currently some fairly low hanging fruit
in that regard.

The edge detection mode is currently configurable in the viewer as:
 "x_only" - only look for horizontal edges
 "y_only" - only look for vertical edges
 "xy" - look for horizontal and vertical edges
 "none" - don't look for any edges

There is a "delete_edges" property (default true) which results in edge
points being squashed to NAN values. It's useful to be able to turn this
off and then it's possible to set the "cloud_mode" to "edges" so that
the debug cloud colors highlight which points would be deleted.