context: track dedicated codebook update timestamp
Since it doesn't seem inherently clear whether we should be updating the
codebook in cases where we failed to track it seems risky to be assuming
that ctx->last_tracking->frame->timestamp corresponds to the time that
the codebook was last updated.

To help have codebook maintenance be a little more orthogonal this adds
a ctx->last_codebook_update_time that we can be more confident about
refering to, regardless of how ctx->last_tracking gets maintained or
whether we skip codebook updates for some tracking attempts.