context: calculate to_ground transform based on gravity vector
If the device doesn't set a valid pose transformation but it does set a
gravity vector then calculate a rotation that simply aligns the gravity
vector with the -Z axis.

The iOS/avf device backend no longer sets a pose.

This simplifies what's required from the device layer and in particular
makes it easier for us to handle Kinect device rotations or for
generating recordings with gravity/orientation information per-frame.