Associate joint inference code with a context ("inferrer") structure
Instead of parsing the JSON_Value joint_map repeatedly at the start of
inference we now parse this once while creating a 'struct
joints_inferrer' which is passed back into the api when we calculate
weights or start inference.

This also namespaces the joint inference code under 'joints_inferrer_'

The plan is to change a bit what data is kept in a joint map, possibly
moving the definition of bones out into a separate file but also
possibly referring to labels by name instead of index values which seems
very susceptible to silent errors if we ever change our labels. Name
based lookups would be a little more costly though so ideally we
wouldn't be repeating it frequently.