context: make codeword member names more self-documenting
I was finding it hard to read code with member names like:

  ->m, ->n, ->ts, ->tl, ->nc

As with the classification symbols I think these were originally based
on the names used in the paper that our motion-based segmentation was
initially based on. It now seems more important to be able to read the
code standalone than in connection with the original paper. In
corresponding order the new member names are:

  ->mean
  ->n
  ->create_timestamp
  ->last_update_timestamp
  ->consecutive_update_count