context: split out resolve of bg codewords + retiring old codewords
Since it was a fairly orthogonal step done before classifying points;
the loop that ranks codewords by ->n to determine the background and
that also handles retiring/pruning old codewords has been factored out
into a separate stage.

This also updates the way old codewords are pruned so the depth-based
sort order is preserved.

This adds some initial support for a debug viewing mode that visualizes
the codebook words instead of the point cloud being processed. (The mode
is controlled via a 'codebook_debug_view' property associated with this
stage)