update representation of bones
Bone topology is now loaded from a more constrained description that
implies we have a single root bone and a directed acyclic graph of
children.

The context now was an array of per-pone info where we can find the
parent and children of any bone index.

Either iterating through all bones as an array or traversing the bones
as a hierarchy from the root is somewhat simpler now and in particular we
don't have lots of nested searches for bones matching given head/tail
joints since per-bone children are tracked explicitly.

While making this change we found that the previous
calc_mismatched_bones() code had a mistake in it's inner loop while
looking for child bones due to using the outer loop's 'i' index value.
So although we now see the skeleton refine step taking more time, that's
likely because the previous version wasn't doing all the work it was
intended to do.

This adds a new debug-only tracking stage that lets us visualize
tracking history updates and also visualize a skeleton prediction
relative to these historic tracking results.

Ideally this would be split out but this patch also updates how target
sequences are represented. Target sequences are now stored in a single
.json file with an array of 'frames' where each frame has an array of
'joints' instead of storing target poses in terms of bones. We now also
track a per-frame timestamp and an 'anchor joint' which can be used
when overlaying a target pose on a tracked skeleton to decide which
joint should match the user's target skeleton.