targets: set neck as fallback anchor joint if unspecified
Since adding support for reading a per-frame anchor joint we didn't
explicitly check whether an anchor joint was specified in json and would
get back a value of zero from the parson api. For pre-existing
target sequences that resulted in the head being used as the anchor
joint which doesn't give good results when mapping to a tracked skeleton
while someone's head is tilted forwards.