context: track per-bone validity status, as we do for joints
If either joint associated with a bone is invalid or if the joints for
the parent of a bone are invalid we can't calculate the bone length and
angle of the bone relative to its parent so we will also consider the
bone as invalid.

The viewer has been updated to consider that gm_skeleton_get_bone() may
return NULL for invalid bones.