recording: Handle IO in separate thread with RAM buffering
While recording we can now specify how much RAM we are prepared to
sacrifice for storing frames that haven't been written to storage yet.

The recording state can be queried so that the viewer can show how much
RAM is currently being used for unwritten frames so it's possible to see
when you will run out of buffer space (at which point the recording will
start to drop frames).

The actual writing is now handled in a separate thread so we don't have
to drop frames if we can't write to storage as fast as we are capturing
new frames from the camera (very likely).

When recording is 'stopped' via the viewer then the UI reflects the fact
that there may still be IO pending to finish writing the results to
storage.