context: allow viewing layers of codebook in debug view
The resolve_background_debug stage now has a 'debug_codebook_layer'
property (default 0) and if set to a value > 0 then when visualizing
the codebook as a point cloud we only show the first N codewords
for each codebook entry.

When projecting the codebook into a point cloud we now also pass the
resolved background codewords so that we can colour those codewords
red.