context: check for NaN values when projecting clusters into depth images
For each candidate cluster we create a depth_image which is first
cleared to a background value defined by our decision trees and then
each point in the cluster is projected into the 2D depth image. We
weren't skipping over NaN values and there was some inconsistency
between running on different hardware whether that would result in NaN
values in the final depth image (in some cases the projected points
would fail the clip test and therefore not be included).

NaN values in the depth images can later result in us crashing if we end
up inferring joint positions with NaN values so adding this early check
ultimately fixes crashes we were seeing when testing on an iPhone X.