context: support merging tiny clusters into large
This is currently enabled by default but so far in practice it's not
clear that it makes a noticeable/significant difference.

The original thought had been that this could help avoid small speckles
of background within the middle of the body due to codebook pollution
but we don't currently have enough test videos to be able to measure the
benefit (or not) of this.