context: Simplify joint position sanitisation and add time window
Add a time window to all sanitisations so that we don't perform
sanitisation based on stale tracking data. Also, simplify joint position
sanitisation. It's been explicitly named to draw attention to the fact
that it's working based on the velocity of each joint over tracking
history, and it now applies to all joints rather than allowing it to work
on just the parent bone. When possible, correct joint positions based on
displacement from a bone head joint rather than on the absolute position.