context: support merging large clusters by bounding box overlap
The hope was that we could handle cases where the body gets broken into
multiple clusters by arm occlusion by grouping clusters based on their
bounding boxes.

In practice this does seem to work in some cases but especially early on
before we've started tracking then this merging will be too greedy and
makes us less likely to track.

For now this is disabled by default and it needs to be refined but in
the meantime it's still useful for debugging/experimentation with paused
frames.

This also updates how the rgb debug images for the clustering stage are
created, to be consistent with how the debug cloud points are coloured.