context: infer with cropped clusters
This removes the projection into 'training camera' coordinates before
running label + joint inference. The assumption now is that the
downsampled intrinsics are a close enough match for what the decision
trees expect.

So instead of having a stage before running inference that creates an
array of depth images from reprojected clusters, we now have a
_CROP_CLUSTER_IMAGE stage
that's part of the same per-cluster loop that will create just one depth
image at a time that is cropped around the cluster currently being
considered.

Now that the depth images used for inference may be smaller than the
full depth buffer we pass in x0,y0 offsets to the joint inference code
that needs to be able to transform coordinates back into 3D points
according to intrinsics that are based on the downsampled (not cropped)
resolution.