context: debug cloud assertions + fixes
Help ensure we don't combine the addition of indexed and non-indexed
debug cloud points.

This also fixes a bug in add_debug_cloud_xyz_from_pcl_xyzl_and_indices()
when resizing the debug_cloud_indices to provide room for the new
indices whereby it was basing the new size on the current size of the
debug_cloud which had just been resized itself to add room for the same
indices.