context: fix project_point_into_codebook()
There was a mistake with the calculation of the y coordinate when
projecting points into the codebook which was resulting in us cropping a
few rows of pixels at the bottom of every frame.

This also avoids some repeated derivation of intrinsics based on the
depth_camera_intrinsics and seg_res (using the recently added
downsampled_intrinsics instead).