joints_inferrer_infer_fast: Fix using points outside of the cluster
Occasionally infer_fast would pick points outside of the cluster as it
just takes the average x/y of each point in the cluster and doesn't verify
that point actually lies in the cluster. We now take the nearest point to
the average x/y instead. I don't think the non-fast alternative has this
issue, but I've not verified that.