context: Add 'codebook_update_delay' property
Add a property that allows us to define a period of time after failed
tracking to stop updating the motion detection codebook. This allows us
to avoid polluting the codebook in the case of a false-negative with
skeleton validation (or in the case that the depth data is valid but the
tracking produced bad results).