device: playback debug messages and remove looping special case
When we have to skip frames to keep up with the progress of time during
recording playback we had a special case to avoid looping since in the
old code we wanted to avoid the complexity of handling looping in
multiple places.

Since the implementation has evolved we can simply remove this special
case and the logic for looping should Just Work (tm).

This fixes an issue where playback would get stuck on the last frame
and repeatedly hit the skip frame logic and then repeatedly hit the
special case that would stop the playback from looping.